Fix integer overflow in get_admin_stats() breaking the whole Stats tab
The heatmap's 30-day window computed 30 * 24 * 60 * 60 * 1000 as plain
32-bit integer math (2,592,000,000), overflowing Postgres's int4 max
(~2.147 billion) and making the entire RPC error out with "22003:
integer out of range" on every call. The app's loadAdminStats() checks
res.ok but doesn't log failures, so this surfaced only as the Stats tab
silently showing zeros for everything, not just the heatmap. Fixed by
casting to bigint. Applied and verified against the live database:
push_subscribers now correctly returns 18, matching the real row count.

+++ supabase/03-device-stats.sql
@@ -98,6 +98,11 @@ as $$
         )
     ),
     'heatmap', (
+      -- The 30-day window below needs the ::bigint cast: 30*24*60*60*1000
+      -- = 2,592,000,000, which overflows Postgres's plain "integer" type
+      -- (max ~2.147 billion) and made the whole function error out with
+      -- "22003: integer out of range" — the 7-day windows above stay
+      -- under that limit so they didn't show the bug.
       select coalesce(json_agg(row_to_json(t)), '[]'::json)
       from (
         select
@@ -106,7 +111,7 @@ as $$
           count(*) as count
         from usage_events
         where type = 'app_open'
-          and ts >= (extract(epoch from now()) * 1000 - 30 * 24 * 60 * 60 * 1000)
+          and ts >= (extract(epoch from now()) * 1000 - 30::bigint * 24 * 60 * 60 * 1000)
         group by dow, hour
       ) t
     )
